pyproject.toml
[tool.poetry]
name = "deep-search-api"
version = "1.0.0"
description = "Deep Search API with FastAPI, Faiss, and OpenAI embeddings"
authors = ["Deep Search Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
openai = "^1.13.0"
faiss-cpu = "^1.8.0"
sqlalchemy = "^2.0.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"
numpy = "^1.26.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.0"
black = "^24.3.0"
pytest = "^8.1.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
